add useTweetstormHandler
