+++ .gitignore
@@ -62,3 +62,4 @@ typings/
 
 package-lock.json
 .env
+.DS_Store
